% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RnaSeqTutorials-package.R
\docType{package}
\name{RnaSeqTutorial03 package}
\alias{RnaSeqTutorialXX}
\alias{RnaSeqTutorialXX-package}
\alias{RnaSeqTutorial03 package}
\title{RnaSeqTutorials package}
\description{
A simple description of the RnaSeqTutorial03 package
}
\section{Tutorials}{

This is the first in a series of tutorials
\itemize{
\item\code{03_differential_gene_expression} a differential gene expression tutorial using DESeq2
}
}

\examples{
	\dontrun{
	learnr::run_tutorial("03_differential_gene_expression", package = "RnaSeqTutorial03")
	}
}
\seealso{
The vignette
}
\author{
Nicolas Delhomme [aut,cre]
}
\keyword{internal}
\keyword{package}
